Move individual specs to one large spec file
